adapt to `gix-features`

+++ gix-index/src/file/verify.rs
@@ -27,7 +27,7 @@ impl File {
             let should_interrupt = AtomicBool::new(false);
             let actual = gix_features::hash::bytes_of_file(
                 &self.path,
-                num_bytes_to_hash as usize,
+                num_bytes_to_hash,
                 checksum.kind(),
                 &mut gix_features::progress::Discard,
                 &should_interrupt,
